requirements.txt
django==2
djangorestframework==3.8.2
gunicorn
psycopg2-binary
requests
django-storages
boto
requests
python-dateutil

responses
ipdb
ipython
coreapi
openapi_codec
drf-yasg
drf-yasg[validation]
# temp:
django-cors-headers
jsonschema
faker